refactor: rename accept to accpets
