--- src/helper/accept/index.ts
@@ -1,1 +1,0 @@
-export { accepts } from './accept'
