+++ src/helper/accepts/index.ts
@@ -1,0 +1,1 @@
+export { accepts } from './accepts'
